Change HashSet to LinkedHashSet in tests
In Java 8 there has been a change in the hashing used for string with
hash-based maps. See:
http://permalink.gmane.org/gmane.comp.java.openjdk.core-libs.devel/10361

This causes tests to fail since we depend on the order of the
HashSet. In this commit we changed our tests to use the LinkedHashSet
that guarantees order and thus is the same in Java 7 and Java 8.