Add plugin-descriptor.properties and update pom.xml
Starting with elasticsearch version 2, plugins are now required to
contain a metadata file called plugin-descriptor.properties in the root
directory of the packaged zip (outside the .jar).

In order to avoid the properties between this new file and pom.xml
getting out of sync, try to share them as much as possible.

The compiler properties are available since maven.compiler.plugin
version 2
https://maven.apache.org/plugins/maven-compiler-plugin/compile-mojo.html#source

More info here:
https://www.elastic.co/guide/en/elasticsearch/plugins/2.3/plugin-authors.html
and here:
https://github.com/elastic/elasticsearch/blob/2.x/dev-tools/src/main/resources/plugin-metadata/plugin-descriptor.properties